Creates a nextflow workflow for guide assignment
This workflow can be configured to use cleanser, umi thresholds, or
scepter to perform guide assignments.

It can also be run standalone or as a sub workflow in a larger pipeline.
