+++ single-cell/guide-assignment/.gitignore
@@ -1,0 +1,4 @@
+__pycache__
+work
+.nextflow*
+.R*
