+++ single-cell/guide-assignment/nextflow_pipeline/example.config
@@ -1,0 +1,3 @@
+params.INPUT_FILE = "/Users/tc325/Documents/CRISPR-JAMBOREE/single-cell/guide-assignment/test_data/gasperini_guide_assignment_input_minimal.h5mu"
+params.OUTPUT_FILE = "/Users/tc325/Documents/CRISPR-JAMBOREE/single-cell/guide-assignment/out.h5mu"
+params.ASSIGNMENT_METHOD = "sceptre"
